\name{annotatebox}
\alias{annotatebox}
\title{Annotate Box
}
\description{Show a box with corner
  points annotated
}
\usage{
annotatebox(Rbox = matrix(ncol = 4, nrow = 8), add = TRUE)
}
\arguments{
  \item{Rbox}{Box matrix
}
  \item{add}{logical, TRUE= add to existing plot
}
}

\value{Graphical Side effects
}

\author{
Jonathan M. Lees<jonathan.lees@unc.edu>
}


\seealso{stress, stressSETUP, annotateplane
}
\examples{

annotatebox()



}
\keyword{misc}
