\name{annotateplane}
\alias{annotateplane}
\title{Annotate Plane
}
\description{Show a box and plane with the three
  points annotated
}
\usage{
annotateplane(Rp = matrix(ncol = 4, nrow = 3), add = TRUE)
}
\arguments{
  \item{Rp}{Points matrix
}
  \item{add}{logical, TRUE= add to existing plot
}
}

\value{Graphical Side effects
}

\author{
Jonathan M. Lees<jonathan.lees@unc.edu>
}


\seealso{stress, stressSETUP, annotatebox
}
\examples{
annotateplane()



}
\keyword{misc}
