\name{BMOD}
\alias{BMOD}
\title{Plot block model for gravity
}
\description{
  Plot block model with several
  polygonal structures for gravity modeling.
}
\usage{
BMOD(bill, nstn = 100, PLOT = TRUE, obs=NULL)
}
\arguments{
  \item{bill}{Polygon data from ModelG
}
  \item{nstn}{number of stations
}
  \item{PLOT}{logical, TRUE=plot
  }
   \item{obs}{Observed Profile Gravity (x,g)
  }  
}
\details{Used internally in interactive modeling.
}
\value{graphical Side effects.
}

\author{
Jonathan M. Lees<jonathan.lees@unc.edu>
}


\seealso{ModelG
}

\keyword{misc}
