\name{DO.HALFSPACE}
\alias{DO.HALFSPACE}
\title{Heat Equation in Halfspace
}
\description{Heat Equation in Halfspace
}
\usage{
DO.HALFSPACE()
}
\details{Several solution for different times are
  displayed in succession.
}
\value{Graphical Side effects
}

\author{
Jonathan M. Lees<jonathan.lees@unc.edu>
}

\seealso{DO.DYKE
}
\examples{

DO.HALFSPACE()


}
\keyword{misc}
