\name{geophys-package}
\alias{geophys-package}
\alias{geophys}
\docType{package}
\title{Geophysics
}
\description{
Geophysics, Continuum Mechanics, Mogi Models, Mohr's Circles, Okada Model
}
\details{
\tabular{ll}{
Package: \tab geophys\cr
Type: \tab Package\cr
Version: \tab 1.2-0\cr
Date: \tab 2010-05-12\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}

}
\author{
Jonathan M. Lees

Maintainer: Jonathan M. Lees<jonathan.lees@unc.edu>
}
\references{
Mogi
}

\keyword{ package }
\seealso{
RSEIS, RFOC, zoeppritz
}
\examples{

\dontrun{

stress()

}



}
