\name{get.brachiopod}
\alias{get.brachiopod}
\title{Brachiopod icon
}
\description{set Brachiopod icon
}
\usage{
get.brachiopod()
}
\details{Program returns the outline of a brachiopod
  for strain analysis.
}
\value{2D Matrix of coordinates.
}

\author{
Jonathan M. Lees<jonathan.lees@unc.edu>
}
\note{Used in conjunction with  draw.brachiopod
}


\seealso{draw.brachiopod
}
\examples{

gb = get.brachiopod()

draw.brachiopod(gb)




}
\keyword{misc}
