\name{PXY}
\Rdversion{1.1}
\alias{PXY}
\docType{data}
\title{
Deformation of a volcano 
}
\description{
Deformation of Pozzuoli Italy 1982
}
\usage{data(PXY)}
\format{
  The format is:
List of 2
 $ x: num [1:75] 0.0237 0.3478 0.5252 0.8689 0.8387 ...
 $ y: num [1:75] 1 0.991 0.975 0.928 0.962 ...
}
\details{Used as example data for  Mogi Source inversion.
  

}
\source{Digitized from publication listed in reference.

}
\references{publication


  Berrino, G., Corrado, G., Luongo, G., and  Toro, B.,
  Ground deformation and gravity changes accompanying the 1982 Pozzuoli
  Uplift, in Monograph:On the 1982-1984 bradyseismic crisis at
  Phlegraean Fields (Italy)
  eds: Barberi, F., Hill, D P., Innocenti, F., Luongo, G., Treuil, M.,
  Bulletin Volcanologique, vol.47, no.2, special issue, pp.187-200, 1984 <doi:https://doi.org/10.1007/BF01961548>.
  

}
\examples{
data(PXY)

plot(PXY)


}
\keyword{datasets}
